backend/requirements.txt
Django>=5.0.1
djangorestframework>=3.14.0
django-cors-headers>=4.3.1
psycopg2-binary>=2.9.9
djoser>=2.2.0
djangorestframework-simplejwt>=5.3.0
openai>=1.0.0,<2.0.0
python-dotenv>=1.0.0
pydantic>=2.10.0
stripe>=7.8.0
Pillow>=10.3.0
